requirements.txt
lightning
streamlit_drawable_canvas
streamlit_webrtc
supervision[assets]
typer[all]
ultralytics
vidgear
yolov5